removed endpoint, updated tests
