+++ domain-cc/cc-app/src/python_src/api.py
@@ -3,20 +3,17 @@ import logging
 import sys
 import time
 from functools import wraps
-from typing import Optional, Tuple
+from typing import Tuple
 
 from fastapi import FastAPI, HTTPException, Request
 
 from .pydantic_models import (
-    Claim,
     ClaimLinkInfo,
     ClassifiedContention,
     ClassifierResponse,
     Contention,
-    PredictedClassification,
     VaGovClaim,
 )
-from .util.brd_classification_codes import get_classification_name
 from .util.logging_dropdown_selections import build_logging_table
 from .util.lookup_table import ContentionTextLookupTable, DiagnosticCodeLookupTable
 from .util.sanitizer import sanitize_log
@@ -67,31 +64,6 @@ def get_health_status():
         raise HTTPException(status_code=500, detail="Lookup table is empty")
 
     return {"status": "ok"}
-
-
-def log_lookup_table_match(
-    classification_code: int,
-    contention_text: str,
-):
-    is_in_dropdown = contention_text.strip().lower() in dropdown_values
-    log_as_json({"is_in_dropdown": sanitize_log(is_in_dropdown)})
-    log_contention_text = contention_text if is_in_dropdown else "Not in dropdown"
-
-    if classification_code:
-        already_mapped_text = contention_text.strip().lower()  # do not leak PII
-        log_as_json({"lookup_table_match": sanitize_log(already_mapped_text)})
-    elif is_in_dropdown:
-        log_as_json(
-            {
-                "lookup_table_match": sanitize_log(
-                    f"No table match for {log_contention_text}"
-                )
-            }
-        )
-    else:
-        log_as_json(
-            {"lookup_table_match": sanitize_log("No table match for free text entry")}
-        )
 
 
 def log_as_json(log: dict):
@@ -188,78 +160,6 @@ def log_contention_stats_decorator(func):
         return result
 
     return wrapper
-
-
-def log_claim_stats(claim: Claim, classification: Optional[PredictedClassification]):
-    if classification:
-        classification_code = classification.classification_code
-        classification_name = classification.classification_name
-    else:
-        classification_code = None
-        classification_name = None
-
-    contention_text = claim.contention_text or ""
-    is_in_dropdown = contention_text.strip().lower() in dropdown_values
-    is_mapped_text = dropdown_lookup_table.get(contention_text, None) is not None
-    log_contention_text = (
-        contention_text if is_mapped_text else "unmapped contention text"
-    )
-
-    log_as_json(
-        {
-            "claim_id": sanitize_log(claim.claim_id),
-            "claim_type": sanitize_log(claim.claim_type),
-            "classification_code": classification_code,
-            "classification_name": classification_name,
-            "contention_text": log_contention_text,
-            "diagnostic_code": sanitize_log(claim.diagnostic_code),
-            "form526_submission_id": sanitize_log(claim.form526_submission_id),
-            "is_in_dropdown": is_in_dropdown,
-            "is_lookup_table_match": classification_code is not None,
-        }
-    )
-
-
-@app.post("/classifier", deprecated=True)
-def get_classification(claim: Claim) -> Optional[PredictedClassification]:
-    """[DEPRECATED] Use /va-gov-claim-classifier instead"""
-    log_as_json(
-        {
-            "claim_id": sanitize_log(claim.claim_id),
-            "form526_submission_id": sanitize_log(claim.form526_submission_id),
-        }
-    )
-    classification_code = None
-    if claim.claim_type == "claim_for_increase":
-        classification_code = dc_lookup_table.get(claim.diagnostic_code)[
-            "classification_code"
-        ]
-
-    if claim.contention_text and not classification_code:
-        classification_code = dropdown_lookup_table.get(claim.contention_text)[
-            "classification_code"
-        ]
-
-    if claim.claim_type == "new":
-        log_lookup_table_match(classification_code, claim.contention_text)
-    else:
-        log_as_json({"diagnostic code": sanitize_log(claim.diagnostic_code)})
-
-    if classification_code:
-        classification_name = get_classification_name(classification_code)
-        classification = {
-            "classification_code": classification_code,
-            "classification_name": classification_name,
-        }
-    else:
-        classification = None
-
-    log_as_json({"classification": classification})
-    log_claim_stats(
-        claim,
-        PredictedClassification(**classification) if classification else None,
-    )
-    return classification
 
 
 @app.post("/claim-linker")
